Translated chitkit app to react component, Now backend use "rails s -p 3002" everything else the same
