--- chatkit/package.json
@@ -1,23 +1,0 @@
-{
-  "name": "react-chat-tutorial",
-  "version": "0.1.0",
-  "private": true,
-  "dependencies": {
-    "@pusher/chatkit": "^0.7.17",
-    "cors": "^2.8.4",
-    "express": "^4.16.3",
-    "pusher-chatkit-server": "^0.12.0",
-    "react": "^16.2.0",
-    "react-dom": "^16.2.0",
-    "react-scripts": "1.1.1"
-  },
-  "devDependencies": {
-    "concurrently": "^3.5.1"
-  },
-  "scripts": {
-    "start": "concurrently \"node ./server.js\" \"react-scripts start\"",
-    "build": "react-scripts build",
-    "test": "react-scripts test --env=jsdom",
-    "eject": "react-scripts eject"
-  }
-}
