+++ client/package.json
@@ -2,7 +2,7 @@
   "name": "pet-share-app",
   "version": "0.1.0",
   "private": true,
-  "proxy": "http://localhost:3000/",
+  "proxy": "http://localhost:3002/",
   "dependencies": {
     "react": "^16.4.2",
     "react-dom": "^16.4.2",
@@ -12,10 +12,17 @@
     "react-router-dom": "^4.3.1",
     "styled-components": "3.4.1",
     "axios": "^0.18.0",
-    "classnames": "^2.2.6"
+    "classnames": "^2.2.6",
+    "@pusher/chatkit": "^0.7.17",
+    "cors": "^2.8.4",
+    "express": "^4.16.3",
+    "pusher-chatkit-server": "^0.12.0"
+  },
+  "devDependencies": {
+    "concurrently": "^3.5.1"
   },
   "scripts": {
-    "start": "react-scripts start",
+    "start": "concurrently \"node ./server.js\" \"react-scripts start\"",
     "build": "react-scripts build",
     "test": "react-scripts test --env=jsdom",
     "eject": "react-scripts eject"
