+++ client/server.js
@@ -1,0 +1,50 @@
+const express = require('express');
+const bodyParser = require('body-parser');
+const cors = require('cors');
+const Chatkit = require('pusher-chatkit-server');
+
+const app = express();
+// app.set('port', process.env.PORT || 3001);
+
+const chatkit = new Chatkit.default({
+  instanceLocator: 'v1:us1:cfee68c8-9258-4fb5-9623-955b2d1b1f35',
+  key: 'feb47bdd-6a9e-436a-bf09-7ea47886fb20:xvEucarGv0vasG0/Wn4J22tPDd//VUZzGvArRaeFYZo=',
+});
+
+app.use(bodyParser.urlencoded({ extended: false }));
+app.use(bodyParser.json());
+app.use(cors());
+
+app.post('/users', (req, res) => {
+  const { username } = req.body;
+  chatkit
+    .createUser({ 
+      id: username, 
+      name: username 
+    })
+    .then(() => res.sendStatus(201))
+    .catch(error => {
+      if (error.error_type === 'services/chatkit/user_already_exists') {
+        res.sendStatus(200);
+      } else {
+        res.status(error.status).json(error);
+      }
+    });
+});
+
+app.post('/authenticate', (req, res) => {
+  const authData = chatkit.authenticate({ userId: req.query.user_id });
+  res.status(authData.status).send(authData.body);
+});
+
+
+const PORT = 3001;
+console.log(PORT);
+app.listen(PORT, err => {
+  if (err) {
+    console.log('here');
+    console.error(err);
+  } else {
+    console.log(`Running on port ${PORT}`);
+  }
+});
