+++ client/src/components/Chatkit/Chatkit.js
@@ -1,0 +1,42 @@
+import React, { Component } from 'react';
+import UsernameForm from './UsernameForm';
+import ChatScreen from './ChatScreen';
+
+class Chatkit extends Component {
+  constructor() {
+    super();
+    this.state = {
+      currentUsername: '',
+      currentScreen: 'WhatIsYourUsernameScreen'
+    };
+    this.onUsernameSubmitted = this.onUsernameSubmitted.bind(this);
+  }
+
+  onUsernameSubmitted(username) {
+    fetch('http://localhost:3001/users', {
+      method: 'POST',
+      headers: {
+        'Content-Type': 'application/json',
+      },
+      body: JSON.stringify({ username }),
+    })
+      .then(response => {
+        this.setState({
+          currentUsername: username,
+          currentScreen: 'ChatScreen'
+        });
+      })
+      .catch(error => console.error('error', error));
+  }
+
+  render() {
+    if (this.state.currentScreen === 'WhatIsYourUsernameScreen') {
+      return <UsernameForm onSubmit={this.onUsernameSubmitted} />;
+    }
+    if (this.state.currentScreen === 'ChatScreen') {
+      return <ChatScreen currentUsername={this.state.currentUsername} />;
+    }
+  }
+}
+
+export default Chatkit;
