+++ client/src/components/Chatkit/TypingIndicator.js
@@ -1,4 +1,4 @@
-import React, { Component } from 'react'
+import React, { Component } from 'react';
 
 class TypingIndicator extends Component {
   render() {
