+++ client/src/components/Chatkit/UsernameForm.js
@@ -1,4 +1,4 @@
-import React, { Component } from 'react'
+import React, { Component } from 'react';
 
 class UsernameForm extends Component {
  constructor(props) {
