+++ client/src/components/Chatkit/WhosOnlineList.js
@@ -1,4 +1,4 @@
-import React, { Component } from 'react'
+import React, { Component } from 'react';
 
 class WhosOnlineList extends Component {
   renderUsers() {
